+++ src/test/java/nextstep/subway/unit/AuthServiceMockTest.java
@@ -3,8 +3,8 @@ package nextstep.subway.unit;
 import nextstep.auth.application.AuthService;
 import nextstep.auth.infra.JwtTokenProvider;
 import nextstep.member.application.MemberService;
-import nextstep.member.application.dto.TokenRequest;
-import nextstep.member.application.dto.TokenResponse;
+import nextstep.auth.application.dto.TokenRequest;
+import nextstep.auth.application.dto.TokenResponse;
 import nextstep.member.domain.Member;
 import nextstep.member.config.exception.PasswordMatchException;
 import org.junit.jupiter.api.DisplayName;
